REF: feature_data_mag
