+++ q2_types/__init__.py
@@ -22,7 +22,6 @@ importlib.import_module('q2_types.sample_data')
 importlib.import_module('q2_types.metadata')
 importlib.import_module('q2_types.multiplexed_sequences')
 importlib.import_module('q2_types.kraken2')
-importlib.import_module('q2_types.feature_data_mag')
 importlib.import_module('q2_types.genome_data')
 importlib.import_module('q2_types.kaiju')
 importlib.import_module('q2_types.reference_db')
